requirements.txt
# ML Platform - PHASE 0 Requirements
# ===================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# HTTP Client (for testing)
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Optional: Development Tools
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.0
# flake8==6.1.0
